Add files via upload
You have been asked to implement the Okta Sign-In Widget for authentication purposes. There are three tasks in this Use Case.
However, before you begin any of the tasks, you must first: 1. Download the filehererename it to index.html and put it by itself in a folder on your local machine. 2. Sign up for an account or Log in to your account at https://www.netlify.com/. 3. Go to Projects > Add a new project > Deploy manually  
￼
  4. Upload the folder to your new project. 
￼
Now that you have your own copy of the project, proceed to Task 1.


Implement a basic sign-in widget without any unnecessary customizations, authenticate to Okta, and redirect the user (Aidan Nadia) to the dashboard.
1) Open the index.html file you downloaded with a text editor or code editor.
2) Perform the tasks necessary to enable the Okta Sign-in Widget.
3) Go back to netlify.com go to Projects > [your project name] > Deploys > Upload your updated local project into project.

￼
4) To log in to your Okta org from your index.html page, you must enable Cross-Origin Resource Sharing (CORS) in Okta.
5) Launch your site by going back to Project overview and clicking on the preview icon to the left of the project name.
￼
6) Log in to your Okta org as Aidan Nadia. If necessary, return to your Glitch project and edit the source code of your index.html file and try again.
7) You can find more information as you work through this task in the following places:
* https://github.com/okta/okta-signin-widget
* https://developer.okta.com/code/javascript/okta_sign-in_widget/
Go to the Task 2 tab to begin Task 2.


Using the Okta console, enable Security Questions as an Authenticator for Recovery in password policy rules.
Create a Factor Enrollment policy with Email being Optional.
Set up a new Authentication Policy and Rule. Name the policy: Interns MFA Policy. Name its associated Rule: Interns MFA Rule.
Apply the policy to the Interns group and to the Okta Dashboard App.
Go to the Task 3 tab to begin Task 3.


Go back to the sign-in widget, authenticate with aidan.nadia@oktane.com/Testme321! but DO NOT enroll Aidan in MFA.
Congratulations! You have completed all the use cases for this exam. Please submit your exam for grading.